requirements.txt
Django==4.0.5
Pillow==9.1.1
djangorestframework==3.13.1